+++ tests/testthat/test_encapsulate.R
@@ -12,6 +12,22 @@ task = tsk("iris")
 learner = lrn("classif.debug")
 learner$param_set$values = list(message_train = 1, warning_train = 1, message_predict = 1, warning_predict = 1)
 
+test_that("encapsulation is automatically enabled", {
+  tmp = lrn("classif.debug")
+  expect_equal(tmp$encapsulate, c(train = "none", predict = "none"))
+  expect_null(get_private(tmp)$.encapsulate)
+
+  tmp$fallback = lrn("classif.featureless")
+  expect_equal(tmp$encapsulate, c(train = "evaluate", predict = "evaluate"))
+  expect_equal(get_private(tmp)$.encapsulate, c(train = "evaluate", predict = "evaluate"))
+
+  tmp = lrn("classif.debug")
+  tmp$encapsulate = c(train = "none", predict = "none")
+  tmp$fallback = lrn("classif.featureless")
+  expect_equal(tmp$encapsulate, c(train = "none", predict = "none"))
+  expect_equal(get_private(tmp)$.encapsulate, c(train = "none", predict = "none"))
+})
+
 test_that("evaluate / single step", {
   row_ids = 1:120
   expect_message(expect_warning(disable_encapsulation(learner)$train(task, row_ids)))