+++ NEWS.md
@@ -1,3 +1,12 @@
+# mlr3 0.13.1-9000
+
+* Setting a fallback learner for a learner with encapsulation in its default
+  settings now automatically sets encapsulation to `"evaluate"` (#763).
+* `as_task_classif()` and `as_task_regr()` now support the construction of tasks
+  using the formula interface, e.g. `as_task_regr(mpg ~ ., data = mtcars)`
+  (#761).
+
+
 # mlr3 0.13.1
 
 * Improved performance for many operations on `ResampleResult` and